requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==5.1.2
django-allauth==65.3.1
django-allauth-ui==1.5.3
django-celery-beat==2.7.0
django-channels==0.7.0
django-environ==0.12.0
django-recaptcha==4.0.0
django-rosetta==0.10.0
django-timezone-field==7.1
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jiter==0.8.2
jwt==1.3.1
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
pillow==10.4.0
polib==1.2.0
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.8.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
service-identity==24.2.0
six==1.17.0
slippers==0.6.2
sniffio==1.3.1
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
